issue-204 레이아웃 기본 완성
