+++ app/src/main/java/com/boostcamp/dreampicker/presentation/feed/voted/VotedViewModel.java
@@ -1,4 +1,4 @@
-package com.boostcamp.dreampicker.presentation.result;
+package com.boostcamp.dreampicker.presentation.feed.voted;
 
 import com.boostcamp.dreampicker.data.local.room.dao.VotedFeedDao;
 import com.boostcamp.dreampicker.data.local.room.entity.VotedFeed;
@@ -9,12 +9,12 @@ import androidx.lifecycle.LiveData;
 import androidx.paging.LivePagedListBuilder;
 import androidx.paging.PagedList;
 
-class ResultViewModel extends BaseViewModel {
-    private static final int PAGE_SIZE = 10;
+public class VotedViewModel extends BaseViewModel {
+    private static final int PAGE_SIZE = 5;
     @NonNull
     private LiveData<PagedList<VotedFeed>> votedFeedList;
 
-    ResultViewModel(@NonNull final VotedFeedDao votedFeedDao) {
+    VotedViewModel(@NonNull final VotedFeedDao votedFeedDao) {
         this.votedFeedList =
                 new LivePagedListBuilder<>(votedFeedDao.selectAll(), PAGE_SIZE).build();
     }
