+++ app/src/main/java/com/boostcamp/dreampicker/presentation/feed/voted/VotedViewModelFactory.java
@@ -1,4 +1,4 @@
-package com.boostcamp.dreampicker.presentation.result;
+package com.boostcamp.dreampicker.presentation.feed.voted;
 
 import com.boostcamp.dreampicker.data.local.room.dao.VotedFeedDao;
 
@@ -6,20 +6,20 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-public class ResultViewModelFactory implements ViewModelProvider.Factory {
+public class VotedViewModelFactory implements ViewModelProvider.Factory {
     @NonNull
     private final VotedFeedDao votedFeedDao;
 
-    ResultViewModelFactory(@NonNull VotedFeedDao votedFeedDao) {
+    public VotedViewModelFactory(@NonNull VotedFeedDao votedFeedDao) {
         this.votedFeedDao = votedFeedDao;
     }
 
     @NonNull
     @Override
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-        if(modelClass.isAssignableFrom(ResultViewModel.class)) {
+        if(modelClass.isAssignableFrom(VotedViewModel.class)) {
             //noinspection unchecked
-            return (T) new ResultViewModel(votedFeedDao);
+            return (T) new VotedViewModel(votedFeedDao);
         } else {
             throw new IllegalArgumentException("ViewModel Not Found");
         }
