+++ app/src/main/res/layout/item_voted_feed.xml
@@ -1,0 +1,77 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout>
+    <data>
+        <variable
+            name="feed"
+            type="com.boostcamp.dreampicker.data.local.room.entity.VotedFeed"/>
+    </data>
+    <androidx.cardview.widget.CardView
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:cardCornerRadius="5dp"
+        android:elevation="3dp"
+        app:cardUseCompatPadding="true">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/voted_feed_container_size_height">
+
+            <ImageView
+                android:id="@+id/iv_voted_feed_profile_image"
+                android:layout_width="@dimen/profile_image_size_large"
+                android:layout_height="@dimen/profile_image_size_large"
+                android:layout_margin="@dimen/space_x_small"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"/>
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="@dimen/space_x_small"
+                android:textSize="@dimen/text_x_large"
+                android:textColor="@color/colorBlack"
+                android:text="윤OO"
+                app:layout_constraintTop_toTopOf="@id/iv_voted_feed_profile_image"
+                app:layout_constraintStart_toEndOf="@id/iv_voted_feed_profile_image"/>
+
+            <TextView
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="@dimen/space_x_small"
+                android:layout_marginTop="8dp"
+                android:textSize="@dimen/text_large"
+                android:ellipsize="marquee"
+                android:focusable="true"
+                android:singleLine="true"
+                android:marqueeRepeatLimit="marquee_forever"
+                app:layout_constraintEnd_toStartOf="@id/iv_voted_feed_image_a"
+                app:layout_constraintStart_toEndOf="@id/iv_voted_feed_profile_image"
+                app:layout_constraintBottom_toBottomOf="@id/iv_voted_feed_profile_image"/>
+
+            <ImageView
+                android:id="@+id/iv_voted_feed_image_a"
+                android:layout_width="@dimen/voted_feed_image_size"
+                android:layout_height="@dimen/voted_feed_image_size"
+                android:layout_marginEnd="@dimen/space_xx_small"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toStartOf="@id/iv_voted_feed_image_b"/>
+
+            <ImageView
+                android:id="@+id/iv_voted_feed_image_b"
+                android:layout_width="@dimen/voted_feed_image_size"
+                android:layout_height="@dimen/voted_feed_image_size"
+                android:layout_marginStart="@dimen/space_x_small"
+                android:scaleType="fitXY"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"/>
+
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </androidx.cardview.widget.CardView>
+
+</layout>
